% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_Tmem68.R
\name{subset_Tmem68}
\alias{subset_Tmem68}
\title{Extract subset chromosomes from Tmem68}
\usage{
subset_Tmem68(object, chr = NULL)
}
\arguments{
\item{object}{list object with mediator, annotation, and other elements}

\item{chr}{chromosome(s) to select (all if \code{NULL})}
}
\description{
Extract subset chromosomes from Tmem68
}
