% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_Tmem68.R
\name{Tmem68_transform}
\alias{Tmem68_transform}
\title{Transform Tmem68 for use in mediation.}
\usage{
Tmem68_transform(Tmem68)
}
\arguments{
\item{Tmem68}{list object with mediator, annotation, and other elements}
}
\description{
This routine was done once on original data.
}
